requirements.txt
argon2-cffi==21.1.0
asgiref==3.5.0
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
django-allauth==0.49.0
django-filter==21.1
django-passwords==0.3.12
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djlint==0.7.5
entrypoints==0.3
fcm-django==1.0.11
filelock==3.6.0
firebase-admin==5.2.0
fonttools==4.28.2
google-api-core==2.7.2
google-api-python-client==2.44.0
google-auth==2.6.4
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.18.2
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.4
oauthlib==3.2.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydotplus==2.0.2
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywin32==301
pywinpty==1.1.4
PyYAML==6.0
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.1
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tomlkit==0.9.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.7.0
